requirements.txt
python-telegram-bot==20.7
apscheduler
python-dotenv
httpx
